ndt7: refactor to isolate download's measurer
This commit moves the download measurer into a separate file. Now
we have a nice API that we can invoke to measure.

This commit is part of a serious of commit that aim to refactor
the download to be a pipeline of goroutines.
